komanda/requirements/base.txt
Django
psycopg2-binary
gunicorn
djangorestframework